2022/12/16

+++ src/main/java/com/sxh/usercenter/Model/domain/Article.java
@@ -24,12 +24,12 @@ public class Article implements Serializable {
     /**
      * 发布人id
      */
-    private Integer userId;
+    private long userId;
 
     /**
      * 队伍Id
      */
-    private Integer teamId;
+    private long teamId;
 
     /**
      * 正文部分
